refactor(claude): isolate TUI composer locks
